Fix longstanding issue with moduleGraphNodes - no hs-boot files case
In the case when we tell moduleGraphNodes to drop hs-boot files the idea
is to collapse hs-boot files into their hs file nodes. In the old code

* nodeDependencies changed edges from IsBoot to NonBoot
* moduleGraphNodes just dropped boot file nodes

The net result is that any dependencies of the hs-boot files themselves
were dropped. The correct thing to do is

* nodeDependencies changes edges from IsBoot to NonBoot
* moduleGraphNodes merges dependencies of IsBoot and NonBoot nodes.

The result is a properly quotiented dependency graph which contains no
hs-boot files nor hs-boot file edges.

Why this didn't cause endless issues when compiling with boot files, we
will never know.