Update documentation of LiberalTypeSynonyms
We no longer require LiberalTypeSynonyms to use 'forall' or an unboxed
tuple in a synonym.

I also removed that kind checking before expanding synonyms "could be changed".
This was true when type synonyms were thought of macros, but with
the extensions such as SAKS or matchability I don't see it changing.