Fix bug in weak loop-breakers in OccurAnal
Note [Weak loop breakers] explains why we need to track variables free
in RHS of rules.  But we need to do this for /inactive/ rules as well
as active ones, unlike the rhs_fv_env stuff.

So we now have two fields in node Details, one for free vars of
active rules, and one for free vars of all rules.

This was shown up by #20820, which is now fixed.