hadrian: Flag cabal flag handling
This patch basically deletes some ad-hoc handling of Cabal Flags and
replaces it with a correct query of the LocalBuildInfo. The flags in the
local build info can be modified by users by passing hadrian options

For example (!4331)

```
*.genapply.cabal.configure.opts += --flags=unregisterised
```

And all the flags specified by the `Cabal Flags` builder were already
passed to configure properly using `--flags`.