Core Lint: collect args through floatable ticks
We were not looking through floatable ticks when collecting arguments in
Core Lint, which caused `checkCanEtaExpand` to fail on something like:

```haskell
reallyUnsafePtrEquality
  = \ @a ->
      (src<loc> reallyUnsafePtrEquality#)
        @Lifted @a @Lifted @a
```

We fix this by using `collectArgsTicks tickishFloatable` instead of
`collectArgs`, to be consistent with the behaviour of eta expansion
outlined in Note [Eta expansion and source notes] in GHC.Core.Opt.Arity.

Fixes #21152.