Add an inline pragma to lookupVarEnv
The containers bump reduced the size of the Data.IntMap.Internal.lookup
function so that it no longer experienced W/W. This means that the size
of lookupVarEnv increased over the inlining threshold and it wasn't
inlined into the hot code path in substTyVar.

See containers#821, #21159 and !7638 for some more explanation.

-------------------------
Metric Decrease:
    LargeRecord
    T12227
    T13386
    T15703
    T18223
    T5030
    T8095
    T9872a
    T9872b
    T9872c
    TcPlugin_RewritePerf
-------------------------