TTG: Make (HsQuote GhcTc) isomorphic to ()
An untyped bracket `HsQuote p` can never be constructed with
`p ~ GhcTc`. This is because we don't typecheck `HsQuote` at all.

That's OK, because we also never use `HsQuote GhcTc`.

To enforce this at the type level we make `HsQuote GhcTc` isomorphic
to () and impossible to construct otherwise, by using TTG field
extensions to make all constructors, except for `XQuote` (which takes `()`),
unconstructable with `DataConCantHappen`