Update user guide example rewrite rules formatting
Change the rewrite rule examples to include a space between the
composition of `f` and `g` in the map rewrite rule examples.

Without this change, if the user has locally enabled the extension
OverloadedRecordDot the copied example will result in a compile time
error that `g` is not a field of `f`.

```
    • Could not deduce (GHC.Records.HasField "g" (a -> b) (a1 -> b))
        arising from selecting the field ‘g’
```