hadrian: Fix split-sections transformer
The splitSections transformer has been broken since -dynamic-too support
was implemented in hadrian. This is because we actually build the
dynamic way when building the dynamic way, so the predicate would always
fail.

The fix is to just always pass `split-sections` even if it doesn't do
anything for a particular way.

Fixes #21138