docs: Fix path to GHC API docs in index.html
In the make bindists we generate documentation in docs/ghc-<VER> but the
hadrian bindists generate docs/ghc/ so the path to the GHC API docs was
wrong in the index.html file.

Rather than make the hadrian and make bindists the same it was easier to
assume that if you're using the mkDocs script that you're using hadrian
bindists.

Fixes #21509

(cherry picked from commit 68d1ea5fe083e0b3ed114bbf02a642c7481ab5d7)