SpecConstr: Don't unbox strict, Boxed args
If the demand on a argument binder is strict and Boxity analysis says that we
use the box, it's probably a bad idea to specialise away the box.
Otherwise, worker/wrapper would have unboxed the arg before.

See Note [Boxity in SpecConstr] for details.

Fixes #21562 and perhaps makes progress towards #20321.